pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-slide-generator"
version = "1.0.0"
description = "AI-powered slide generator from text to Google Slides and HTML presentations"
authors = [{name = "AI Assistant", email = "ai@assistant.com"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.8"
keywords = ["slides", "presentation", "google-slides", "html", "ai", "automation"]

dependencies = [
    "click>=8.0.0",
    "google-api-python-client>=2.100.0",
    "google-auth>=2.20.0",
    "google-auth-oauthlib>=1.0.0",
    "google-auth-httplib2>=0.1.0",
    "markdown>=3.5.0",
    "pyyaml>=6.0.0",
    "beautifulsoup4>=4.12.0",
    "jinja2>=3.1.0",
    "python-dotenv>=1.0.0",
    "colorama>=0.4.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]

[project.scripts]
slide-generator = "slide_generator.cli:main"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=src/slide_generator --cov-report=term-missing"